# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RT Management API",
        "description": "API documentation for RT Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rumah": {
            "get": {
                "tags": [
                    "Rumah"
                ],
                "summary": "Get list of rumah",
                "operationId": "12e1d89f91088a37ce297c59e870f5a4",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rumah"
                ],
                "summary": "Create new rumah",
                "operationId": "de75c51f0ec5ac405e65f9b033f61272",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nomor_rumah",
                                    "status"
                                ],
                                "properties": {
                                    "nomor_rumah": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/rumah/{id}": {
            "get": {
                "tags": [
                    "Rumah"
                ],
                "summary": "Get detail rumah",
                "operationId": "bce69ff8f6ef8c720e6d5bbf3b0d3487",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rumah"
                ],
                "summary": "Update rumah",
                "operationId": "5c3bdc736f369b24ad728eee700aa69d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nomor_rumah": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rumah"
                ],
                "summary": "Delete rumah",
                "operationId": "00655f14624b51e5779dd47997bf6a0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                }
            }
        },
...
    },
...
}